split pattern.py into submodules
